Move login from main to login activity
